+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+### Fixed
+- In config, the channel range field is now used by the daemon.
+
+### Changed
+- In config, channel range fields is now a list of floats (`[min, max]`).
+
 ## [2022.11.0]
 
 ### Added